+++ src/kfactory/typings.py
@@ -16,14 +16,18 @@ from klayout import lay
 from typing_extensions import TypeAliasType
 
 if TYPE_CHECKING:
-    from collections.abc import Callable
+    from collections.abc import Callable, Iterable
 
-    from .decorators import WrappedKCellFunc, WrappedVKCellFunc
+    from cachetools import Cache
+
+    from .conf import CheckInstances
+    from .decorators import PortsDefinition, WrappedKCellFunc, WrappedVKCellFunc
     from .instance import ProtoInstance, ProtoTInstance
     from .kcell import BaseKCell, ProtoKCell, ProtoTKCell, VKCell
     from .layer import LayerEnum
     from .pin import ProtoPin
     from .port import ProtoPort
+    from .schematic import TSchematic
 
 T = TypeVar("T")
 K = TypeVar("K", bound="ProtoKCell[Any, Any]")
@@ -157,3 +161,26 @@ KCellSpec: TypeAlias = (
     "int | str | KCellSpecDict | ProtoTKCell[Any] | Callable[..., ProtoTKCell[Any]]"
 )
 AnyCellSpec: TypeAlias = "int | str | KCellSpecDict | ProtoTKCell[Any] | VKCell | Callable[..., ProtoTKCell[Any]] | Callable[..., VKCell]"  # noqa: E501
+
+
+class CellKwargs(TypedDict, total=False):
+    set_settings: bool
+    set_name: bool
+    check_ports: bool
+    check_pins: bool
+    check_instances: CheckInstances
+    snap_ports: bool
+    add_port_layers: bool
+    cache: Cache[int, Any] | dict[int, Any]
+    basename: str
+    drop_params: list[str]
+    register_factory: bool
+    overwrite_existing: bool
+    layout_cache: bool
+    info: dict[str, MetaData]
+    post_process: Iterable[Callable[[ProtoTKCell[Any]], None]]
+    debug_names: bool
+    tags: list[str]
+    lvs_equivalent_ports: list[list[str]]
+    ports: PortsDefinition
+    schematic_function: Callable[..., TSchematic[Any]]
